fixed stat gathering overflow and improved copy speeds

+++ pftool.h
@@ -11,7 +11,7 @@ void manager(int rank, struct options o, int nproc, path_list *input_queue_head,
 void manager_workdone(int rank, int sending_rank, int *proc_status);
 int manager_add_paths(int rank, int sending_rank, path_list **queue_head, path_list **queue_tail, int *queue_count);
 int manager_add_buffs(int rank, int sending_rank, work_buf_list **workbuflist, int *workbufsize);
-void manager_add_copy_stats(int rank, int sending_rank, int *num_copied_files, int *num_copied_bytes);
+void manager_add_copy_stats(int rank, int sending_rank, int *num_copied_files, double *num_copied_bytes);
 void manager_add_examined_stats(int rank, int sending_rank, int *num_examined);
 
 //worker rank operations
